renderQuery in test utils
